IB/hfi1: change krcvqs module parameter type from byte to uint
There is a bug in 3.10.0-123.el7.x86_64 which causes param_get_byte
to return the wrong value. This causes krcvqs parameter to display
incorrectly in sysfs. The workaround is to set the param type as uint.

INTERNAL: PR 130000 - Driver not reporting value of module parameter, krcvqs

+++ hfi.h
@@ -1650,7 +1650,7 @@ extern unsigned int hfi1_cu;
 extern unsigned int user_credit_return_threshold;
 extern uint num_rcv_contexts;
 extern unsigned n_krcvqs;
-extern u8 krcvqs[];
+extern uint krcvqs[];
 extern int krcvqsset;
 extern uint kdeth_qp;
 extern uint loopback;
